// src/generated/models/aks_designmastersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface aks_designmasters {
  aks_autonumber: string;
  aks_designcode?: string;
  aks_designimage?: string;
  aks_designimage_timestamp?: number;
  aks_designimage_url?: string;
  aks_designimageid?: string;
  aks_designmasterid: string;
  aks_designno: string;
  aks_itemname: string;
  aks_itemnameformulabased?: string;
  aks_itemnamename?: string;
  aks_parentdesign?: string;
  aks_prefix?: string;
  aks_supplierid: string;
  aks_supplieridname?: string;
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: Date;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: number;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: Date;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: Date;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: string;
  statecodename?: string;
  statuscode?: string;
  statuscodename?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
}